вынес код карты в отдельный файл
